requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
apimatic-core==0.2.15
apimatic-core-interfaces==0.1.5
apimatic-requests-client-adapter==0.1.6
attrs==25.3.0
blinker==1.9.0
boto3==1.38.15
botocore==1.38.15
CacheControl==0.12.14
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
dash==3.0.4
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
faiss-cpu==1.11.0
Flask==3.0.3
frozenlist==1.6.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpickle==3.0.4
jsonpointer==2.4
langchain==0.3.25
langchain-anthropic==0.3.6
langchain-community==0.3.24
langchain-core==0.3.59
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langsmith==0.3.42
MarkupSafe==3.0.2
marshmallow==3.26.1
msgpack==1.1.0
multidict==6.4.3
mypy_extensions==1.1.0
narwhals==1.39.0
nest-asyncio==1.6.0
numpy==2.2.5
openai==1.78.1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
plotly==6.1.0
propcache==0.3.1
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
s3transfer==0.12.0
setuptools==80.3.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
squareup==38.1.0.20240919
stripe==12.1.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.0.6
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
streamlit==1.41.1
mailchimp-marketing==3.0.80
beautifulsoup4==4.12.3
python-pptx==0.6.23
kaleido==0.2.1  # For plotly static image export
